Fix login, sign-up not showing any errors on bad inputs
This has to do with devise and turbo not playing nicely with each other.

On a bad submit the page doesn't redirect again to itself cause turbo
loads the page from cache or somethin, BODERLINE is we are unable to redirect
to the same login/sign-up page after a bad input and hence are unable to see
the error messages for our bad inputs which come usually from flash[:error]

Fix: there are 2 suggested in the members-only project in rails, here i
went with the 1st one where i generate devise views and disable turbo for
all forms.